Fix for storing the dta-File map position. This affects only files of size > 2GB for which the map position previously was not stored correctly. Stata 13+ reads the map position and checks if the tags match.
The information is stored as a uint64_t, which is not present in C++. We stored a wrong value. Files with a size >2GB written before this fix are unable to be read by Stata. Importing them with readstata13 is not affected because we do not use the map position for importing.